feat: back to home button always visible
