+++ src/components/home/SidePanel.vue
@@ -12,11 +12,11 @@ function toggleInformationPanel() {
   <div class="flex shadow-sm min-h-screen">
     <div
       class="px-6 py-8 bg-white flex flex-col gap-2 w-[450px] max-h-screen overflow-y-auto scrollbar-hide"
-      :class="{ hidden: !panelStore.isInformationPanelShown }"
+      :class="{ invisible: !panelStore.isInformationPanelShown }"
     >
       <slot></slot>
     </div>
-    <div class="relative max-h-0 top-28">
+    <div class="max-h-0 top-28">
       <UiPanelControlButton
         :anchor-position="'left'"
         :is-open="panelStore.isInformationPanelShown"
