+++ src/views/StationView.vue
@@ -5,10 +5,7 @@ import { useMapStore } from '@/stores/map'
 import { useViewsStore } from '@/stores/views'
 import { useLayersStore } from '@/stores/layers'
 import { useLineViewsStore, useStationViewsStore } from '@/stores/views'
-import router from '@/router'
-import ChevronArrowLeft from '@/assets/icons/chevron-left.svg'
 import UiStationHeader from '@/components/ui/UiStationHeader.vue'
-import UiButton from '@/components/ui/UiButton.vue'
 import { apiClientService } from '@/services/api.client'
 import FooterArea from '@/components/home/FooterArea.vue'
 import type { LineModel } from '@/model/lines.model'
@@ -16,6 +13,7 @@ import type { StationModel } from '@/model/stations.model'
 import { viewList } from '@/model/views.model'
 import ServicesStation from '@/components/station/ServicesStation.vue'
 import PointsOfInterestsStation from '@/components/station/PointsOfInterestsStation.vue'
+import BackButton from '@/components/home/BackButton.vue'
 
 const mapStore = useMapStore()
 const viewStore = useViewsStore()
@@ -60,24 +58,12 @@ onMounted(async () => {
     poi: true,
   })
 })
-
-function backButtonClicked() {
-  const line = lineNumber.value.toString()
-  router.push('/line/' + line)
-  mapStore.viewPoint = `line${line}`
-  stationStore.emptySelectedStation()
-}
 </script>
 
 <template>
   <div class="flex flex-col items-start py-0 gap-2">
     <div class="flex items-center p-0 gap-4">
-      <UiButton
-        class="shadow-md rounded-lg p-2 flex gap-2.5 shrink-0 grow-0"
-        @click="backButtonClicked"
-      >
-        <img :src="ChevronArrowLeft" />
-      </UiButton>
+      <BackButton></BackButton>
       <UiStationHeader
         v-if="state.lineDescription && state.stationDescription"
         :line="state.lineDescription?.id!"
